refactor(userPage): change users page structure New user interface file, content of the tabs was separated and the object of the tabs was changed
