+++ src/pages/privileges/outlets/users/config/usersTabs.config.jsx
@@ -1,14 +1,14 @@
-const privilegeUserTabsConfig = [
-  {
+const privilegeUserTabsConfig = {
+  privilegesUsers: {
     id: "privileges-users",
     isDisabled: false,
     label: "Users",
   },
-  {
+  privilegesInvitations: {
     id: "privileges-invitations",
     isDisabled: false,
     label: "Invitations",
   },
-];
+};
 
 export { privilegeUserTabsConfig };
